feat: feat: implementar capa de servicios para la gestión de detalles de ventas

+++ src/main/java/com/gahelrodriguez/kinalapp/repository/ProductoRepository.java
@@ -2,9 +2,9 @@ package com.gahelrodriguez.kinalapp.repository;
 
 import com.gahelrodriguez.kinalapp.entity.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.List;
 
 public interface ProductoRepository extends JpaRepository<Producto, Long> {
-    List<Producto> findByEstado(int estado);
+
+    List<Producto> findByEstado(Long estado);
 }
